Group code and move to dedicated files in main package.
